mcp_run POC
